Static Buffer (#149)
- Creates a static buffer of completed messages. These aren't refreshed
in Ink so it should be more performant
- Moves all messages through the SSE stream so its a source of truth
- Combines partial messages together using an id
- Moves in progress messages to the static buffer when they are
completed

## Todo
We need to chunk in progress messages to 5 lines so that you can scroll
up while things are actively updating. When the rendering output is
larger than the screen it jumps to the bottom when new updates come in.
5 lines is a sanity number that is probably smaller than most terminal
heights, so rendering won't cause a jump!