Hallucination is enabled only for llm agents (#402)
### Disable Hallucination Validation for Non-LLM Agents

In this update, we've disabled hallucination validation for non-LLM
agents. The primary reason for this change is that non-LLM agents do not
have structured prompts for source tagging, which is essential for the
validation process. As such, the current hallucination validation
mechanism is not applicable for these agents.

Additionally, non-LLM agents are evaluated with a more careful, manual
process, ensuring their outputs are reliable through comprehensive
evaluation.

### Key Focus on LLM Agents:

The critical adjustment here is the emphasis on evaluating LLM agents in
real-time. LLM agents operate as single-part agents with dynamically
generated flows, and as such, require more focused evaluation during
execution. Given that these agents don’t undergo traditional evaluation
testing, the on-the-fly evaluation will be critical in ensuring output
quality and mitigating hallucinations in real-time.

---------

Co-authored-by: Lukasz Jagiello <lukasz@tempest.team>
Co-authored-by: Michał Gryko <michal@tempest.team>
Co-authored-by: Eric Skram <eric@tempest.team>