Faster research agent (#403)
## Summary

The original research agent was fundamentally a 'dumb' pipeline which
processed searches sequentially. It had no ability to estimate the size
of the research task or stop when the question was answered.

The new architecture uses a supervisor-worker pattern inspired by Open
Deep Research. The new research agent is comprised of 3 main components:

_Lead Researcher_ (Supervisor)

- Analyzes the user's query to select research depth (quick, standard,
deep)
- Chunks the user's query into distinct topics and delegates them to
sub-researchers
- Determines, based on summaries returned by the sub-researchers,
whether the question has been answered

_Sub-Researchers_ (Workers)

- Execute searches and extractions using Tavily
- Summarizes query results and offloads them to an in-memory (not in LLM
context) store
- Returns a synthesis to the lead researcher

_Report Writer_

- Executes after the lead researcher determines that the question has
been answered
- Reads the summaries generated by all sub-researchers and can elect to
read raw results if it would be helpful
- Writes the final report and returns it

---

Please read
https://github.com/tempestteam/atlas/blob/fast-research/packages/bundled-agents/src/research/README.md
for a full explanation of the new agent's architecture.