Disable memory prompt enrichment (#453)
## Ticket number

## Contents
1. This PR disables injection of memories in user prompts. When working
memory was allocated for a conversation, it was being injected verbatim
into the 'user's' message since the entire input is treated as a user
message in the conversation agent.

Since memories are also loaded inline in the conversation agent
[[1](https://github.com/tempestteam/atlas/blob/main/packages/system/agents/conversation/conversation.agent.ts#L274)]
then this was duplicative anyway.

| The problem |
| --- |
| <img width="1344" height="758" alt="image"
src="https://github.com/user-attachments/assets/8e866f1e-b051-4848-857a-6de868357f1f"
/> |


---

The _correct_ longer-term solution in this case is to introduce a [Chat
primitive](https://ai-sdk.dev/docs/ai-sdk-ui/overview), whereby user
messages are persisted from the client rather than inserted based on the
prompt provided to the agent.