Recent Conversations (#558)
- Adds recent conversations to the main chat page with support for
loading them and continuing a conversation
- Names conversations using an LLM with at the 1 and 3 response level
(see below)
- Does not try to rename chats that are already named

### Chat naming
Chats are named after the first response from the LLM–to ensure all
chats have a name–and after the third response. The reason for the
second naming is to differentiate chats.

If you started every chat with hello, the name of each chat wouldn't be
very unique. We can continue to tweak this moving forward.

---------

Co-authored-by: Eric Skram <eric@tempest.team>