Buff CC agent output tokens; loosen artifact length limits (#607)
## Contents

This addresses a couple issues identified in my demo today:

1. Increases the maximum output tokens from the Claude Code agent from
default 1k to 30k. This dramatically increases the quality of output
from that agent so it's not arbitrarily constrained.
2. Increases the maximum validation length for artifact summaries. LLMs
are instructed to constrain the summary to ~500 characters, but
sometimes they're off by a few. Widening this constraint means that even
if it misses by XX characters, it'll still pass validation.