Sidebar Chats (#884)
Shows all conversations in the sidebar. Also refetches chat history when
messages return from the daemon. This is important to update the sidebar
when a new conversation is initiated.

A follow up issue for next week has been created to delete
conversations.
https://linear.app/tempestteam/issue/TEM-3342/delete-conversations

Once the share feature from ticket 3331 is completed, this should also
be available from a chat in the sidebar as well

---------

Co-authored-by: Lukasz Jagiello <lukasz@tempest.team>